Verify e2e device identity after connect
