+++ rpc/type.go
@@ -8,13 +8,37 @@ import (
 	"net"
 	"time"
 
+	"github.com/diodechain/diode_go_client/blockquick"
+	"github.com/diodechain/diode_go_client/crypto"
+	"github.com/diodechain/diode_go_client/db"
 	"github.com/diodechain/diode_go_client/edge"
 	"github.com/diodechain/diode_go_client/util"
 )
 
-var (
-	NullData = []byte("null")
+const (
+	confirmationSize  = 6
+	windowSize        = 100
+	rpcCallRetryTimes = 2
+
+	lvbnKey = "lvbn3"
+	lvbhKey = "lvbh3"
 )
+
+var (
+	NullData       = []byte("null")
+	bq             *blockquick.Window
+	enqueueTimeout = 100 * time.Millisecond
+)
+
+type Call struct {
+	id         uint64
+	method     string
+	retryTimes int
+	response   chan interface{}
+	signal     chan Signal
+	data       []byte
+	Parse      func(buffer []byte) (interface{}, error)
+}
 
 // Address represents an Ethereum address
 type Address = util.Address
@@ -64,4 +88,33 @@ func isOpError(netErr error) (isOpError bool) {
 		isOpError = true
 	}
 	return
+}
+
+// LastValid returns the last valid block number and block header
+func LastValid() (uint64, crypto.Sha3) {
+	if bq == nil {
+		return restoreLastValid()
+	}
+	return bq.Last()
+}
+
+func restoreLastValid() (uint64, crypto.Sha3) {
+	lvbn, err := db.DB.Get(lvbnKey)
+	var lvbh []byte
+	if err == nil {
+		lvbnNum := util.DecodeBytesToUint(lvbn)
+		lvbh, err = db.DB.Get(lvbhKey)
+		if err == nil {
+			var hash [32]byte
+			copy(hash[:], lvbh)
+			return lvbnNum, hash
+		}
+	}
+	return 500, [32]byte{0, 0, 91, 137, 111, 20, 109, 80, 251, 76, 143, 80, 134, 152, 142, 201, 98, 250, 205, 7, 108, 135, 20, 235, 135, 65, 44, 186, 4, 161, 71, 238}
+}
+
+func storeLastValid() {
+	lvbn, lvbh := LastValid()
+	db.DB.Put(lvbnKey, util.DecodeUintToBytes(lvbn))
+	db.DB.Put(lvbhKey, lvbh[:])
 }
